myapp/api/requirements.txt
aniso8601==4.1.0
astroid==2.1.0
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
colorama==0.3.9
coverage==4.5.2
Django==1.11.7
docopt==0.6.2
Flask==1.0.2
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gunicorn==19.9.0
idna==2.8
isort==4.2.15
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
more-itertools==5.0.0
mysqlclient==1.3.12
nose==1.3.7
passlib==1.7.1
Pillow==5.1.0
pipreqs==0.4.9
pluggy==0.8.1
psycopg2==2.6.2
py==1.7.0
PyJWT==1.7.1
pylint==2.2.2
pytest==4.1.1
pytz==2018.9
requests==2.21.0
six==1.12.0
SQLAlchemy==1.2.18
typed-ast==1.3.1
urllib3==1.24.1
virtualenv==15.1.0
virtualenvwrapper-win==1.2.4
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.1
yarg==0.1.9
